+++ ConvAppServer/Models/Comment.cs
@@ -8,7 +8,7 @@ namespace ConvAppServer.Models
     public class Comment : Feedbackable
     {
         public int ParentId { get; set; }
-        public int ParentType { get; set; }
+        public byte ParentType { get; set; }
 
         public int CreatorId { get; set; }
         public string Text { get; set; }
